Twiddit_social_ms/requirements.txt
asgiref==3.6.0
Django==4.1.7 
djangorestframework==3.14.0
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
pika